requirements.txt
django==5.2.5
djangorestframework==3.16.1
typesense==1.1.1
pdfminer.six==20240706
